* java compile script optimised

+++ sh-scripts/DistroSourceCommand.fn.sh
@@ -16,79 +16,47 @@ fi
 # Runs DistroCommand by compiling it's source code to a temporary folder 
 #
 DistroSourceCommand(){
-	local OUTPUT_PATH="${MDSC_OUTPUT:-$MMDAPP/output}"
+	local MDSC_OUTPUT="${MDSC_OUTPUT:-$MMDAPP/output}"
 	local MDSC_SOURCE="${MDSC_SOURCE:-$MMDAPP/source}"
 	local MDSC_CACHED="${MDSC_CACHED:-$MMDAPP/output/cached}"
 
-	local DIR_OUT="${MDSC_CACHED:-$MMDAPP/output/cached}/myx/myx.distro-source"
-	local DIR_SRC="${MDSC_SOURCE:-$MMDAPP/source}/myx/myx.distro-source"
+	local DIR_OUT="$MDSC_CACHED/myx/myx.distro-source"
+	local DIR_SRC="$MDSC_SOURCE/myx/myx.distro-source"
 
-	local FIL_OUT="$DIR_OUT/bin/ru/myx/distro/DistroSourceCommand.class"
-	local FIL_CPL="$DIR_SRC/bin/ru/myx/distro/DistroSourceCommand.class"
-	local FIL_SRC="$DIR_SRC/java/ru/myx/distro/DistroSourceCommand.java"
-	
-	local HAS_OUT=$(test -f "$FIL_OUT" && echo 'true' || true)
-	local HAS_CPL=$(test -f "$FIL_CPL" && echo 'true' || true)
-	local HAS_SRC=$(test -f "$FIL_SRC" && echo 'true' || true)
+	set -e
 
-	local USE_OUT=$( test -n "$HAS_OUT" && \
-			test -z "$HAS_CPL" -o ! "$FIL_OUT" -ot "$FIL_CPL" && \
-			test -z "$HAS_SRC" -o ! "$FIL_OUT" -ot "$FIL_SRC" && \
-			echo "true" || true \
-		)
-
-	local USE_CPL=$( test -n "$HAS_CPL" -a -z "$USE_OUT" && \
-			test -z "$HAS_OUT" -o ! "$FIL_CPL" -ot "$FIL_OUT" && \
-			test -z "$HAS_SRC" -o ! "$FIL_CPL" -ot "$FIL_SRC" && \
-			echo "true" || true \
-		)
-
-	local USE_SRC=$( test -n "$HAS_SRC" -a -z "$USE_OUT" -a -z "$USE_CPL" && \
-			test -z "$HAS_OUT" -o ! "$FIL_SRC" -ot "$FIL_OUT" && \
-			test -z "$HAS_CPL" -o ! "$FIL_SRC" -ot "$FIL_CPL" && \
-			echo "true" || true \
-		)
-
-	test -n "$USE_SRC" -o -n "$USE_CPL" -o -n "$USE_OUT" || \
-		( echo "No sources available, need to fetch!" ; false )
-
-	test -z "$USE_OUT" || \
+	if [ -f "$DIR_OUT/bin/ru/myx/distro/DistroSourceCommand.class" ] ; then
 		java -cp "$DIR_OUT/bin" ru.myx.distro.DistroSourceCommand \
-			--output-root "$OUTPUT_PATH" \
+			--output-root "$MDSC_OUTPUT" \
 			--source-root "$MDSC_SOURCE" \
 			--cached-root "$MDSC_CACHED" \
 			"$@"
+		return 0
+	fi
 	
-	test -z "$USE_CPL" || \
+	if [ -f "$DIR_SRC/bin/ru/myx/distro/DistroSourceCommand.class" ] ; then
 		java -cp "$DIR_SRC/bin" ru.myx.distro.DistroSourceCommand \
-			--output-root "$OUTPUT_PATH" \
+			--output-root "$MDSC_OUTPUT" \
 			--source-root "$MDSC_SOURCE" \
 			--cached-root "$MDSC_CACHED" \
 			"$@"
-
-	test -z "$USE_SRC" || (
-		. "$MMDAPP/source/myx/myx.distro-source/sh-lib/RunJavaClassClean.include" ;
+		return 0
+	fi
+	
+	if [ -f "$DIR_SRC/java/ru/myx/distro/DistroSourceCommand.java" ] ; then
+		. "$MMDAPP/source/myx/myx.distro-source/sh-lib/RunJavaClassClean.include"
 		RunJavaClassClean \
 			"myx/myx.distro-source" \
 			ru.myx.distro.DistroSourceCommand \
-			--output-root "$OUTPUT_PATH" \
+			--output-root "$MDSC_OUTPUT" \
 			--source-root "$MDSC_SOURCE" \
 			--cached-root "$MDSC_CACHED" \
-			"$@" \
-	)
-
-	return 0
+			"$@"
+		return 0
+	fi
 	
-	test -z "$TMP_SRC" || (
-		. "$MMDAPP/source/myx/myx.distro-source/sh-lib/RunJavaClassTemp.include" ;
-		RunJavaClassTemp \
-			"$MMDAPP/source/myx/myx.distro-source/java" \
-			ru.myx.distro.DistroSourceCommand \
-			--output-root "$OUTPUT_PATH" \
-			--source-root "$MDSC_SOURCE" \
-			--cached-root "$MDSC_CACHED" \
-			"$@" \
-	)
+	echo "DistroSourceCommand: No sources available, need to fetch!" >&2
+	return 1
 }
